bookings_service/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
PyJWT
passlib[bcrypt]
requests
memory_profiler
pika